Try build for linux-aarch64 from source

+++ recipes/manta/build.sh
@@ -1,13 +1,23 @@
 #!/bin/bash
 set -eu
 
-outdir=$PREFIX/share/$PKG_NAME-$PKG_VERSION-$PKG_BUILDNUM
-mkdir -p $outdir
-mkdir -p $PREFIX/bin
+if [[ $(arch) == "aarch64" ]]; then
+    wget https://github.com/Illumina/manta/releases/download/v$PKG_VERSION/manta-$PKG_VERSION.release_src.tar.bz2
+    tar xvf manta-$PKG_VERSION.release_src.tar.bz2
+    cd manta-$PKG_VERSION.release_src
+    ./configure --prefix=$PREFIX
+    make
+    make install
+    
+else
+    outdir=$PREFIX/share/$PKG_NAME-$PKG_VERSION-$PKG_BUILDNUM
+    mkdir -p $outdir
+    mkdir -p $PREFIX/bin
 
-cp -r * $outdir
-rm -rf $outdir/share/demo
-sed -i.bak 's/__file__/os.path.realpath(__file__)/' $outdir/bin/configManta.py
-ln -s $outdir/bin/configManta.py $PREFIX/bin
-ln -s $outdir/libexec/convertInversion.py $PREFIX/bin
-ln -s $outdir/libexec/denovo_scoring.py $PREFIX/bin
+    cp -r * $outdir
+    rm -rf $outdir/share/demo
+    sed -i.bak 's/__file__/os.path.realpath(__file__)/' $outdir/bin/configManta.py
+    ln -s $outdir/bin/configManta.py $PREFIX/bin
+    ln -s $outdir/libexec/convertInversion.py $PREFIX/bin
+    ln -s $outdir/libexec/denovo_scoring.py $PREFIX/bin
+fi
